requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
multitasking==0.0.9
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.9
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
yfinance==0.1.59
